requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
httpx==0.28.1
PyYAML==6.0.2
efinance>=0.5.0
akshare>=1.12.0
baostock>=0.8.9
termcolor>=2.4.0

